integrate egui_http_native and egui_http_web into epi_http
