+++ Cargo.toml
@@ -3,8 +3,7 @@ members = [
     "egui_demo_app",
     "egui_demo_lib",
     "egui_glium",
-    "egui_http_native",
-    "egui_http_web",
+    "epi_http",
     "egui_web",
     "egui",
     "emath",
