+++ egui_glium/Cargo.toml
@@ -24,7 +24,7 @@ glium = "0.29"
 webbrowser = "0.5"
 
 # feature "http":
-egui_http_native ={path="../egui_http_native",optional=true,features=["http"]}
+epi_http ={path="../epi_http",optional=true}
 
 # feature "persistence":
 directories-next = { version = "2", optional = true }
@@ -36,7 +36,7 @@ chrono = { version = "0.4", optional = true }
 
 [features]
 default = []
-http = ["egui_http_native"]
+http = ["epi_http/native"]
 persistence = [
     "directories-next",
     "egui/persistence",
