--- egui_http_native/Cargo.toml
@@ -1,24 +1,0 @@
-[package]
-name = "egui_http_native"
-version = "0.10.0"
-authors = ["Emil Ernerfeldt <emil.ernerfeldt@gmail.com>"]
-description = "Bindings for using egui natively with http request"
-edition = "2018"
-homepage = "https://github.com/emilk/egui"
-license = "MIT OR Apache-2.0"
-readme = "README.md"
-repository = "https://github.com/emilk/egui"
-categories = ["gui", "game-development"]
-keywords = ["ureq", "egui", "gui", "gamedev"]
-include = [
-    "**/*.rs",
-    "Cargo.toml",
-]
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-[features]
-http=["ureq","epi/http"]
-[dependencies]
-ureq ={version="2",optional=true}
-epi  = {version="0.10.0",path="../epi",optional=true}
-
